% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_samle_paths.R
\name{plot_sample_paths}
\alias{plot_sample_paths}
\title{Plot some sample paths}
\usage{
plot_sample_paths(nint, strike, spot, barrier, vol, rf, maturity, nreps = 20)
}
\arguments{
\item{nint}{number of points on path}

\item{strike}{strike price}

\item{spot}{spot price at pricing moment}

\item{barrier}{continuous barrier}

\item{vol}{volatility}

\item{rf}{risk free rate}

\item{maturity}{time to maturity in years}

\item{nreps}{number of paths to generate}
}
\description{
Plot some sample paths
}
